Refactored simulation dir moved to NP arrays
