+++ main.py
@@ -1,9 +1,10 @@
 import csv
-from utils import *
+from time import time
+
 from core import *
-from sim.sim import run_simulation, run_multiple_simulations_for_average
 from core.units import *
-from config.constants import GameSettings
+from simulation.sim import run_multiple_simulations_for_average
+
 
 def main():
     run_count = 2000
@@ -25,7 +26,7 @@ def main():
             round(geq_sim.models_killed / run_count, 1)
         ])
 
-    attacker_1 = [(custodian_guard, 5)]
+    attacker_1 = [(custodian_guard, 1)]
     meq_sim_1 = run_multiple_simulations_for_average(run_count, Scenario(attacker_1, (meq, 50)))
     ork_sim_1 = run_multiple_simulations_for_average(run_count, Scenario(attacker_1, (oeq, 50)))
     teq_sim_1 = run_multiple_simulations_for_average(run_count, Scenario(attacker_1, (teq, 50)))
@@ -33,7 +34,7 @@ def main():
     geq_sim_1 = run_multiple_simulations_for_average(run_count, Scenario(attacker_1, (geq, 50)))
     record_results("Custodian Guard", meq_sim_1, ork_sim_1, teq_sim_1, veq_sim_1, geq_sim_1)
 
-    attacker_2 = [(custodian_guard_sustained, 5)]
+    attacker_2 = [(custodian_guard_sustained, 1)]
     meq_sim_2 = run_multiple_simulations_for_average(run_count, Scenario(attacker_2, (meq, 50)))
     ork_sim_2 = run_multiple_simulations_for_average(run_count, Scenario(attacker_2, (oeq, 50)))
     teq_sim_2 = run_multiple_simulations_for_average(run_count, Scenario(attacker_2, (teq, 50)))
@@ -41,7 +42,7 @@ def main():
     geq_sim_2 = run_multiple_simulations_for_average(run_count, Scenario(attacker_2, (geq, 50)))
     record_results("Custodian Guard with Sustained Hits", meq_sim_2, ork_sim_2, teq_sim_2, veq_sim_2, geq_sim_2)
 
-    attacker_3 = [(custodian_guard_lethal, 5)]
+    attacker_3 = [(custodian_guard_lethal, 1)]
     meq_sim_3 = run_multiple_simulations_for_average(run_count, Scenario(attacker_3, (meq, 50)))
     ork_sim_3 = run_multiple_simulations_for_average(run_count, Scenario(attacker_3, (oeq, 50)))
     teq_sim_3 = run_multiple_simulations_for_average(run_count, Scenario(attacker_3, (teq, 50)))
@@ -49,16 +50,19 @@ def main():
     geq_sim_3 = run_multiple_simulations_for_average(run_count, Scenario(attacker_3, (geq, 50)))
     record_results("Custodian Guard with Lethal Hits", meq_sim_3, ork_sim_3, teq_sim_3, veq_sim_3, geq_sim_3)
 
-    attacker_4 = [(custodian_guard_lethal_and_sustained, 5)]
+    attacker_4 = [(custodian_guard_lethal_and_sustained, 1)]
     meq_sim_4 = run_multiple_simulations_for_average(run_count, Scenario(attacker_4, (meq, 50)))
     ork_sim_4 = run_multiple_simulations_for_average(run_count, Scenario(attacker_4, (oeq, 50)))
     teq_sim_4 = run_multiple_simulations_for_average(run_count, Scenario(attacker_4, (teq, 50)))
     veq_sim_4 = run_multiple_simulations_for_average(run_count, Scenario(attacker_4, (veq, 50)))
     geq_sim_4 = run_multiple_simulations_for_average(run_count, Scenario(attacker_4, (geq, 50)))
-    record_results("Custodian Guard with Lethal Hits and Sustained Hits", meq_sim_4, ork_sim_4, teq_sim_4, veq_sim_4, geq_sim_4)
+    record_results("Custodian Guard with Lethal Hits and Sustained Hits", meq_sim_4, ork_sim_4, teq_sim_4, veq_sim_4,
+                   geq_sim_4)
 
     with open('simulation_results.csv', 'w', newline='') as csvfile:
-        fieldnames = ['Attacker', 'Space Marines Damage', 'Space Marines Killed', 'Orks Damage', 'Orks Killed', 'Terminators Damage', 'Terminators Killed', 'Vehicles Damage', 'Vehicles Killed', 'Guard Damage', 'Guard Killed']
+        fieldnames = ['Attacker', 'Space Marines Damage', 'Space Marines Killed', 'Orks Damage', 'Orks Killed',
+                      'Terminators Damage', 'Terminators Killed', 'Vehicles Damage', 'Vehicles Killed', 'Guard Damage',
+                      'Guard Killed']
         writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
 
         writer.writeheader()
@@ -77,5 +81,8 @@ def main():
                 'Guard Killed': result[10]
             })
 
+
 if __name__ == "__main__":
+    start_time = time()
     main()
+    print(f"Simulation took {time() - start_time} seconds")
