+++ utils/calculations.py
@@ -1,3 +1,6 @@
+import numpy as np
+from numpy import bool_
+
 from utils.dice import roll
 
 
@@ -35,39 +38,14 @@ def calculate_fnp_damage(
     return wounds_taken, fractional_wound
 
 
-def calculate_models_killed(defender_model_wounds, wound_damage_list):
-    models_killed = 0
-    current_model_index = 0
-    total_models = len(defender_model_wounds)
-
-    for damage in wound_damage_list:
-        if current_model_index >= total_models:
-            break
-
-        current_model_wounds = defender_model_wounds[current_model_index]
-
-        if damage >= current_model_wounds:
-            # Kill the model and move to the next
-            defender_model_wounds[current_model_index] = 0
-            models_killed += 1
-            current_model_index += 1
-        else:
-            # Damage the current model
-            defender_model_wounds[current_model_index] -= damage
-
-    return models_killed, defender_model_wounds
+def count_success(rolls: np.ndarray, threshold: int, modifier: int = 0) -> bool_:
+    if modifier != 0:
+        return np.sum(rolls + modifier >= threshold)
+    return np.sum(rolls >= threshold)
 
 
-def count_success(rolls: list[int], threshold: int, modifier: int = 0) -> int:
-    success_count = 0
-    for r in rolls:
-        if dice_check(r, threshold, modifier):
-            success_count += 1
-    return success_count
-
-
-def count_equal_value_in_list(rolls: list[int], value: int) -> int:
-    return sum(1 for r in rolls if r == value)
+def count_equal_value_in_list(rolls: np.ndarray, value: int) -> int:
+    return np.sum(rolls == value)
 
 
 def dice_check(r: int, threshold: int, modifier: int = 0) -> bool:
