+++ utils/dice.py
@@ -1,19 +1,23 @@
 import random
 
+import numpy as np
+
 
 def roll(sides: int = 6):
     return random.randint(1, sides)
 
 
-def rollx(n: int, sides: int = 6) -> list[int]:
+def rolln(n: int, sides: int = 6) -> np.ndarray:
     """
-    Rolls a die n times and returns a list of the results.
+    Rolls a die n times and returns an array of the results.
     :param n: The number of times to roll the die.
     :param sides: The count of sides on the die. (Default 6)
-    :return: A list of the results of the die rolls.
+    :return: An array of the results of the die rolls.
     """
-    return [roll(sides) for _ in range(n)]
+    return np.random.randint(1, sides + 1, n)
 
 
-def reroll_1s(rolls: list[int]) -> list[int]:
-    return [roll() if r == 1 else r for r in rolls]
+def reroll_1s(rolls: np.ndarray) -> np.ndarray:
+    reroll_mask = (rolls == 1)
+    rolls[reroll_mask] = np.random.randint(1, rolls.max() + 1, reroll_mask.sum())
+    return rolls
